web/requirements.txt
Flask>=2.3.0
gunicorn>=21.0.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
openpyxl>=3.1.0
fpdf2>=2.7.0
